requirements.txt
aiogram>=3.4,<4.0
SQLAlchemy>=2.0,<3.0
aiosqlite>=0.19
asyncpg>=0.29
python-dotenv>=1.0
cryptography>=41.0
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
jinja2>=3.1,<4.0
python-multipart>=0.0.9
